ec/lenovo: Add support for MEC1653 EC
Add support for the MEC1653 EC as used by the Thinkpad T480/480s.

These boards will still use the PMH7 and H8 EC, but not the thermal
management as it causes issues apparently, so disable it when the
MEC1653 EC option is selected. Also enable WiFi and BT as these
options are always present and cannot use GPIO status to determine
their presence.

Change-Id: If82a7d27eb3163f51565c0c6e60cab60753611a7
Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>