change to the t480 memory init params file without spd read pins
the t480s has some pins specified as MEMORYID. they are the ones used to
communicate SPD <-> soldered ram.

the t480 doesn't have this, instead it has 2 dimms and only has dynamic
dimm spd reading logic for 2 slots.

for now the no-gpio approach is safer to copy, probably i'll have to
remove one of the slots and only use the 0x51 one (as it were when
trying to probe it on DTS)

that is until i really figure out how to dump the SPDs and which PCH
CPU straps are used to grab those binaries

Signed-off-by: Filip Lewiński <filip.lewinski@3mdeb.com>